requirements.txt
langchain-google-genai
langgraph
python-dotenv
pydantic
wikipedia
ddgs